[index] move optimize and purge request to datastore module
