+++ core/com.b2international.snowowl.server.console/src/com/b2international/snowowl/server/console/MaintenanceCommandProvider.java
@@ -42,13 +42,13 @@ import com.b2international.snowowl.datastore.cdo.ICDORepositoryManager;
 import com.b2international.snowowl.datastore.commitinfo.CommitInfo;
 import com.b2international.snowowl.datastore.commitinfo.CommitInfoDocument;
 import com.b2international.snowowl.datastore.request.RepositoryRequests;
+import com.b2international.snowowl.datastore.request.repository.OptimizeRequest;
+import com.b2international.snowowl.datastore.request.repository.PurgeRequest;
 import com.b2international.snowowl.datastore.request.repository.RepositorySearchRequestBuilder;
 import com.b2international.snowowl.datastore.server.ServerDbUtils;
 import com.b2international.snowowl.datastore.server.migrate.MigrateRequest;
 import com.b2international.snowowl.datastore.server.migrate.MigrateRequestBuilder;
 import com.b2international.snowowl.datastore.server.migrate.MigrationResult;
-import com.b2international.snowowl.datastore.server.reindex.OptimizeRequest;
-import com.b2international.snowowl.datastore.server.reindex.PurgeRequest;
 import com.b2international.snowowl.datastore.server.reindex.ReindexRequest;
 import com.b2international.snowowl.datastore.server.reindex.ReindexRequestBuilder;
 import com.b2international.snowowl.datastore.server.reindex.ReindexResult;
